projects/p4_advanced_rag/requirements.txt
# === Core RAG Dependencies ===
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0

# === Document Parsing ===
pypdf>=4.0.0
unstructured>=0.15.0
markdown-it-py>=3.0.0

# === Embeddings ===
sentence-transformers>=3.0.0

# === Vector Store ===
chromadb>=0.5.0
faiss-cpu>=1.8.0

# === Sparse Retrieval ===
rank-bm25>=0.2.2

# === Reranking ===
# Cross-encoder reranking via sentence-transformers (already included above)

# === LLM API ===
openai>=1.40.0

# === Evaluation ===
ragas>=0.2.0
datasets>=2.20.0

# === Utilities ===
python-dotenv>=1.0.0
rich>=13.0.0
pydantic>=2.0.0
